Fix an issue that somewhat doesn't exist
ok yes I know that file extensions on Linux don't matter, BUT it does help to identify the type of file without using the file command or opening up the file contents. One could argue the .sh extension means shell script, which is technically true, however it's more likely that people would assume .sh means it runs in the /bin/sh shell, but the top of each file clearly specifies the executing binary is infact /bin/bash, therefore the .bash extension is mopre appropriate.

Also the files weren't marked as executable so anyone cloning the repo wouldn't be able to use ./ and whatever to execute them.